pyproject.toml
[project]
name = "youtubechatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
    "tiktoken>=0.12.0",
    "youtube-transcript-api>=1.2.4",
]
